* First draft at door prizes.

+++ games/views.py
@@ -5,7 +5,7 @@ from django.views.generic import ListView, View
 
 from .models import DoorPrizeWinner, DoorPrizePool
 from event.models import Schedule
-from afol.models import ScheduleAttendee
+from afol.models import ScheduleAttendee, Attendee
 from .forms import DoorPrizeWinnerForm
 
 
@@ -13,10 +13,12 @@ from .forms import DoorPrizeWinnerForm
 class DoorPrizeWinnerView(ListView):
     model = DoorPrizeWinner
 
+
 @method_decorator(staff_member_required, name='dispatch')
 class PickDoorPrizeSchedule(ListView):
     template_name = 'games/doorprize_schedule_list.html'
     model = Schedule
+
 
 @method_decorator(staff_member_required, name='dispatch')
 class PickDoorPrizePersonView(View):
@@ -25,25 +27,29 @@ class PickDoorPrizePersonView(View):
     template_name = 'games/doorprizewinner.html'
 
     def get(self, request, *args, **kwargs):
-        int_number = DoorPrizePool.objects.filter(schedule=kwargs['schedule']).count()
+        obj_scheduled_event = Schedule.objects.get(id=kwargs['schedule'])
+        int_number = DoorPrizePool.objects.filter(schedule=obj_scheduled_event).count()
+        # import ipdb; ipdb.set_trace()
+        # No one is in the group so add people form the list of attendees
         if int_number == 0:
-            obj_fols = ScheduleAttendee.objects.filter(schedule=kwargs['schedule'])
+            obj_past_winners = DoorPrizeWinner.objects.filter(event=obj_scheduled_event.event).values_list('fan',
+                                                                                                           flat=True)
+            obj_fols = ScheduleAttendee.objects.exclude(fan__in=obj_past_winners). \
+                filter(schedule=kwargs['schedule'],
+                       fan__attendee__role=Attendee.ROLE_ALLACCESS,
+                       )
             for fol in obj_fols:
                 obj_fol, created = DoorPrizePool.objects.get_or_create(schedule=fol.schedule, fan=fol.fan)
-        obj_winner = DoorPrizePool.objects.order_by('?').first()
-        obj_entree, created = DoorPrizeWinner.objects.get_or_create(fan=obj_winner.fan, event=obj_winner.schedule.event)
-        obj_winner.delete()
-        # print(obj_winner)
-        #
-        # form = DoorPrizeWinnerForm()
-        # fan = obj_winner.fan
-        # event = obj_winner.schedule.event
-        return render(request, self.template_name, {'number': int_number, 'winner': obj_winner, 'schedule': obj_winner.schedule})
-
+        if DoorPrizePool.objects.all().count():
+            obj_winner = DoorPrizePool.objects.order_by('?').first()
+            obj_entree = DoorPrizeWinner.objects.create(fan=obj_winner.fan, event=obj_winner.schedule.event)
+            obj_winner.delete()
+            return render(request, self.template_name,
+                          {'number': int_number, 'winner': obj_winner, 'schedule': obj_winner.schedule})
+        return render(request, self.template_name,
+                      {'number': int_number, 'winner': None, 'schedule': obj_scheduled_event})
     # def post(self, request, schedule):
     #     form = self.form_class(request.POST)
     #     if form.is_valid():
     #         form.save()
     #     return render(request, self.template_name, {'form': form})
-
-
